+++ src/permifrost/__init__.py
@@ -1,2 +1,2 @@
 # Managed by bumpversion
-__version__ = "0.2.0"
+__version__ = "0.2.1"